+++ crypto/ergvein-crypto.cabal
@@ -23,12 +23,14 @@ library
   hs-source-dirs:      src
   exposed-modules:
     Ergvein.Crypto
+    Ergvein.Crypto.Address
     Ergvein.Crypto.Constants
     Ergvein.Crypto.Keys
     Ergvein.Crypto.WordLists
     Ergvein.Crypto.AES256
   build-depends:
       base                      >= 4.9      && <5
+    , aeson                     >= 1.3      && < 1.5
     , bytestring                >= 0.10     && < 0.11
     , bytestring-trie           >= 0.2.5    && < 0.3
     , cryptonite                >= 0.25     && < 0.26
