+++ crypto/src/Ergvein/Crypto.hs
@@ -6,3 +6,4 @@ import Ergvein.Crypto.Keys as X
 import Ergvein.Crypto.Constants as X
 import Ergvein.Crypto.WordLists as X
 import Ergvein.Crypto.AES256 as X
+import Ergvein.Crypto.Address as X
